Remove fp32 deps from llaflow

+++ ptlflow/models/llaflow/corr.py
@@ -29,8 +29,12 @@ class CorrBlock:
         out_pyramid = []
         for i in range(self.num_levels):
             corr = self.corr_pyramid[i]
-            dx = torch.linspace(-r, r, 2 * r + 1, device=coords.device)
-            dy = torch.linspace(-r, r, 2 * r + 1, device=coords.device)
+            dx = torch.linspace(
+                -r, r, 2 * r + 1, dtype=coords.dtype, device=coords.device
+            )
+            dy = torch.linspace(
+                -r, r, 2 * r + 1, dtype=coords.dtype, device=coords.device
+            )
             delta = torch.stack(torch.meshgrid(dy, dx), axis=-1)
 
             centroid_lvl = coords.reshape(batch * h1 * w1, 1, 1, 2) / 2**i
@@ -42,7 +46,7 @@ class CorrBlock:
             out_pyramid.append(corr)
 
         out = torch.cat(out_pyramid, dim=-1)
-        return out.permute(0, 3, 1, 2).contiguous().float()
+        return out.permute(0, 3, 1, 2).contiguous()
 
     @staticmethod
     def corr(fmap1, fmap2):
@@ -52,4 +56,4 @@ class CorrBlock:
 
         corr = torch.matmul(fmap1.transpose(1, 2), fmap2)
         corr = corr.view(batch, ht, wd, 1, ht, wd)
-        return corr / torch.sqrt(torch.tensor(dim).float())
+        return corr / torch.sqrt(torch.tensor(dim))
